Сделано задание 5-1
