+++ tsconfig.json
@@ -105,5 +105,11 @@
     /* Completeness */
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
-  }
+
+  },
+
+  "exclude": [
+    "vendor"
+  ]
 }
+
